refactor(labels): move compute func into separate package
