+++ pkg/core/resources/labels/suite_test.go
@@ -1,0 +1,11 @@
+package labels_test
+
+import (
+	"testing"
+
+	"github.com/kumahq/kuma/v2/pkg/test"
+)
+
+func TestLabels(t *testing.T) {
+	test.RunSpecs(t, "Labels Suite")
+}
