+++ pkg/core/resources/model/resource.go
@@ -3,7 +3,6 @@ package model
 import (
 	"fmt"
 	"hash/fnv"
-	"maps"
 	"reflect"
 	"strings"
 	"time"
@@ -15,9 +14,6 @@ import (
 	common_api "github.com/kumahq/kuma/v2/api/common/v1alpha1"
 	mesh_proto "github.com/kumahq/kuma/v2/api/mesh/v1alpha1"
 	config_core "github.com/kumahq/kuma/v2/pkg/config/core"
-	model_labels "github.com/kumahq/kuma/v2/pkg/core/resources/model/labels"
-	"github.com/kumahq/kuma/v2/pkg/plugins/runtime/k8s/metadata"
-	"github.com/kumahq/kuma/v2/pkg/util/pointer"
 )
 
 const (
@@ -511,229 +507,6 @@ func resourceOrigin(labels map[string]string) (mesh_proto.ResourceOrigin, bool) 
 		return mesh_proto.ResourceOrigin(labels[mesh_proto.ResourceOriginLabel]), true
 	}
 	return "", false
-}
-
-// Namespace type allows to avoid carrying both 'namespace' and 'systemNamespace' around the code base
-// and depend on this type instead
-type Namespace struct {
-	value  string
-	system bool
-}
-
-var UnsetNamespace = Namespace{}
-
-func NewNamespace(value string, system bool) Namespace {
-	return Namespace{
-		value:  value,
-		system: system,
-	}
-}
-
-func GetNamespace(rm ResourceMeta, systemNamespace string) Namespace {
-	if ns, ok := rm.GetNameExtensions()[K8sNamespaceComponent]; ok && ns != "" {
-		return Namespace{
-			value:  ns,
-			system: ns == systemNamespace,
-		}
-	}
-	return UnsetNamespace
-}
-
-type LabelsOptions struct {
-	Mode           config_core.CpMode
-	IsK8s          bool
-	ZoneName       string
-	Namespace      Namespace
-	ServiceAccount string
-	Workload       string
-}
-
-type LabelsOptionsFunc func(*LabelsOptions)
-
-func NewLabelsOptions(fs ...LabelsOptionsFunc) *LabelsOptions {
-	opts := &LabelsOptions{}
-	for _, f := range fs {
-		f(opts)
-	}
-	return opts
-}
-
-func WithK8s(k8s bool) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.IsK8s = k8s
-	}
-}
-
-func WithNamespace(namespace Namespace) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.Namespace = namespace
-	}
-}
-
-func WithServiceAccount(name string) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.ServiceAccount = name
-	}
-}
-
-func WithWorkload(name string) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.Workload = name
-	}
-}
-
-func WithZone(name string) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.ZoneName = name
-	}
-}
-
-func WithMode(mode config_core.CpMode) LabelsOptionsFunc {
-	return func(opts *LabelsOptions) {
-		opts.Mode = mode
-	}
-}
-
-// ComputeLabels computes labels for a resource based on its type, spec, existing labels, namespace, mesh, mode, k8s and localZone.
-// Only use set / setIfNotExist to set labels as it makes sure the label is on the list of computed labels (that is used in another project).
-func ComputeLabels(
-	rd ResourceTypeDescriptor,
-	spec ResourceSpec,
-	existingLabels map[string]string,
-	mesh string,
-	opts ...LabelsOptionsFunc,
-) (map[string]string, error) {
-	labelsOpts := NewLabelsOptions(opts...)
-	labels := map[string]string{}
-	if len(existingLabels) > 0 {
-		labels = maps.Clone(existingLabels)
-	}
-
-	set := func(k, v string) {
-		if _, ok := model_labels.AllComputedLabels[k]; !ok {
-			panic(fmt.Sprintf("label %q is not in the list of computed labels, update AllComputedLabels list as it is used in another project", k))
-		}
-		labels[k] = v
-	}
-
-	setIfNotExist := func(k, v string) {
-		if _, ok := labels[k]; !ok {
-			set(k, v)
-		}
-	}
-
-	getMeshOrDefault := func() string {
-		if mesh != "" {
-			return mesh
-		}
-		return DefaultMesh
-	}
-
-	if rd.Scope == ScopeMesh {
-		setIfNotExist(metadata.KumaMeshLabel, getMeshOrDefault())
-	}
-
-	if labelsOpts.Mode == config_core.Zone {
-		// If resource can't be created on Zone (like Mesh), there is no point in adding
-		// 'kuma.io/zone', 'kuma.io/origin' and 'kuma.io/env' labels even if the zone is non-federated
-		if rd.KDSFlags.Has(ProvidedByZoneFlag) {
-			setIfNotExist(mesh_proto.ResourceOriginLabel, string(mesh_proto.ZoneResourceOrigin))
-			if labels[mesh_proto.ResourceOriginLabel] != string(mesh_proto.GlobalResourceOrigin) {
-				setIfNotExist(mesh_proto.ZoneTag, labelsOpts.ZoneName)
-				env := mesh_proto.UniversalEnvironment
-				if labelsOpts.IsK8s {
-					env = mesh_proto.KubernetesEnvironment
-				}
-				setIfNotExist(mesh_proto.EnvTag, env)
-			}
-		}
-	}
-
-	if labelsOpts.Namespace.value != "" && labelsOpts.IsK8s && IsLocallyOriginated(labelsOpts.Mode, labels) {
-		setIfNotExist(mesh_proto.KubeNamespaceTag, labelsOpts.Namespace.value)
-	}
-
-	if labelsOpts.Namespace.value != "" && rd.IsPolicy && rd.IsPluginOriginated && IsLocallyOriginated(labelsOpts.Mode, labels) {
-		role, err := ComputePolicyRole(spec.(Policy), labelsOpts.Namespace)
-		if err != nil {
-			return nil, err
-		}
-		set(mesh_proto.PolicyRoleLabel, string(role))
-	}
-
-	if rd.IsProxy {
-		proxy, ok := spec.(ProxyResource)
-		if ok {
-			set(mesh_proto.ProxyTypeLabel, strings.ToLower(string(proxy.GetProxyType())))
-		}
-	}
-
-	if labelsOpts.IsK8s {
-		if labelsOpts.ServiceAccount != "" {
-			set(metadata.KumaServiceAccount, labelsOpts.ServiceAccount)
-		}
-	}
-
-	if labelsOpts.Workload != "" {
-		set(metadata.KumaWorkload, labelsOpts.Workload)
-	}
-
-	return labels, nil
-}
-
-func ComputePolicyRole(p Policy, ns Namespace) (mesh_proto.PolicyRole, error) {
-	if ns.system || ns == UnsetNamespace {
-		// on Universal the value is always empty
-		return mesh_proto.SystemPolicyRole, nil
-	}
-
-	hasTo := false
-	if pwtl, ok := p.(PolicyWithToList); ok && len(pwtl.GetToList()) > 0 {
-		hasTo = true
-	}
-
-	hasFrom := false
-	if pwfl, ok := p.(PolicyWithFromList); ok && len(pwfl.GetFromList()) > 0 {
-		hasFrom = true
-	}
-
-	if hasFrom && hasTo {
-		return "", errors.New("it's not allowed to mix 'to' and 'from' arrays in the same policy")
-	}
-
-	if hasFrom || (!hasTo && !hasFrom) {
-		// if there is 'from' or neither (single item)
-		return mesh_proto.WorkloadOwnerPolicyRole, nil
-	}
-
-	hasSameOrOmittedNamespace := func(tr common_api.TargetRef) bool {
-		return pointer.Deref(tr.Namespace) == "" || pointer.Deref(tr.Namespace) == ns.value
-	}
-
-	isProducerItem := func(tr common_api.TargetRef) bool {
-		switch tr.Kind {
-		case common_api.MeshService, common_api.MeshHTTPRoute:
-			return pointer.Deref(tr.Name) != "" && hasSameOrOmittedNamespace(tr)
-		default:
-			return false
-		}
-	}
-
-	producerItems := 0
-	for _, item := range p.(PolicyWithToList).GetToList() {
-		if isProducerItem(item.GetTargetRef()) {
-			producerItems++
-		}
-	}
-
-	switch {
-	case producerItems == len(p.(PolicyWithToList).GetToList()):
-		return mesh_proto.ProducerPolicyRole, nil
-	case producerItems == 0:
-		return mesh_proto.ConsumerPolicyRole, nil
-	default:
-		return "", errors.New("it's not allowed to mix producer and consumer items in the same policy")
-	}
 }
 
 func PolicyRole(rm ResourceMeta) mesh_proto.PolicyRole {
